move lint config specific defs to new file
