+++ src/api/lint_config.ts
@@ -1,0 +1,13 @@
+import * as abaplint from "@abaplint/core";
+
+export interface GetDefaultConfigOutput {
+  config: abaplint.Config;
+}
+
+const defaultConf: GetDefaultConfigOutput = {
+  config: abaplint.Config.getDefault()
+};
+
+export function getDefaultConfig(): GetDefaultConfigOutput {
+  return defaultConf;
+}
